+++ src/Readable/Chunks.php
@@ -11,10 +11,14 @@ use Innmind\Immutable\{
     Either,
 };
 
+/**
+ * @template-covariant T of LowLevelStream
+ */
 final class Chunks
 {
+    /** @var T */
     private LowLevelStream $stream;
-    /** @var callable(LowLevelStream): Maybe<LowLevelStream> */
+    /** @var callable(T): Maybe<T> */
     private $ready;
     /** @var Maybe<Str\Encoding> */
     private Maybe $encoding;
@@ -24,7 +28,8 @@ final class Chunks
     /**
      * @psalm-mutation-free
      *
-     * @param callable(LowLevelStream): Maybe<LowLevelStream> $ready
+     * @param T $stream
+     * @param callable(T): Maybe<T> $ready
      * @param Maybe<Str\Encoding> $encoding
      * @param positive-int $size
      */
@@ -43,10 +48,14 @@ final class Chunks
     /**
      * @psalm-mutation-free
      * @internal
+     * @template A of LowLevelStream
      *
-     * @param callable(LowLevelStream): Maybe<LowLevelStream> $ready
+     * @param A $stream
+     * @param callable(A): Maybe<A> $ready
      * @param Maybe<Str\Encoding> $encoding
      * @param positive-int $size
+     *
+     * @return self<A>
      */
     public static function of(
         LowLevelStream $stream,
@@ -101,6 +110,9 @@ final class Chunks
         return $finished->flatMap(static fn($fold) => $fold->maybe());
     }
 
+    /**
+     * @psalm-mutation-free
+     */
     public function lazy(): Chunks\Lazy
     {
         return Chunks\Lazy::of(