+++ src/Readable/Lines/Lazy.php
@@ -11,22 +11,27 @@ use Innmind\Immutable\{
     Maybe,
 };
 
+/**
+ * @template-covariant T of LowLevelStream
+ */
 final class Lazy
 {
+    /** @var T */
     private LowLevelStream $stream;
-    /** @var callable(LowLevelStream): Maybe<LowLevelStream> */
+    /** @var callable(T): Maybe<T> */
     private $ready;
     /** @var Maybe<Str\Encoding> */
     private Maybe $encoding;
-    /** @var callable(LowLevelStream): void */
+    /** @var callable(T): void */
     private $rewind;
 
     /**
      * @psalm-mutation-free
      *
-     * @param callable(LowLevelStream): Maybe<LowLevelStream> $ready
+     * @param T $stream
+     * @param callable(T): Maybe<T> $ready
      * @param Maybe<Str\Encoding> $encoding
-     * @param callable(LowLevelStream): void $rewind
+     * @param callable(T): void $rewind
      */
     private function __construct(
         LowLevelStream $stream,
@@ -43,8 +48,10 @@ final class Lazy
     /**
      * @psalm-mutation-free
      * @internal
+     * @template A of LowLevelStream
      *
-     * @param callable(LowLevelStream): Maybe<LowLevelStream> $ready
+     * @param A $stream
+     * @param callable(A): Maybe<A> $ready
      * @param Maybe<Str\Encoding> $encoding
      */
     public static function of(
@@ -55,6 +62,9 @@ final class Lazy
         return new self($stream, $ready, $encoding, static fn() => null);
     }
 
+    /**
+     * @psalm-mutation-free
+     */
     public function rewindable(): self
     {
         return new self(
@@ -69,6 +79,8 @@ final class Lazy
     }
 
     /**
+     * @psalm-mutation-free
+     *
      * @return Sequence<Str>
      */
     public function sequence(): Sequence